backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# PDF Processing
PyMuPDF==1.24.13
python-docx==1.1.2
pdfplumber==0.11.4
Pillow==11.0.0
pytesseract==0.3.13
easyocr==1.7.0

# AI/ML
google-generativeai==0.8.5
openai==1.57.4
sentence-transformers==3.3.1
numpy==1.26.4
opencv-python-headless==4.10.0.84

# Vector & Cache
pinecone>=4.0.0
redis==5.2.1
celery[redis]==5.4.0

# Utilities
python-magic==0.4.27
python-dotenv==1.0.1
python-multipart==0.0.17
aiofiles==24.1.0
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
sse-starlette==2.2.1
tqdm==4.67.1

# Security
cryptography==43.0.3
PyJWT==2.10.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
